% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\docType{data}
\name{feature}
\alias{feature}
\title{Feature}
\format{
A data frame with 50 rows and 2 columns. The column names
are as follows:
\describe{
  \item{Symbol}{Official Gene Symbol}
  \item{ID}{Entrez Gene ID}
}
}
\source{
The 50 gene features were selected through SD.
}
\usage{
feature
}
\description{
The features are made up of 50 genes.
}
\keyword{datasets}
