% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpv_pre.R
\name{hpv_pre}
\alias{hpv_pre}
\title{Prediction of human papillomavirus-positive cervical cancers status}
\usage{
hpv_pre(input_data)
}
\arguments{
\item{input_data}{Tumor samples to be predicted. (Note: Gene expression values of the samples need to be normalized (scaling to [0, 1] range) by the function \link{data_pre} first)}
}
\value{
A dataframe with 2 columns:
\item{Sample_ID}{Tumor samples to be predicted.}
\item{HPV_status}{Prediction results ("1" indicates HPV+G1, and "0" HPV+G2).}
}
\description{
\code{hpv_pre} Prediction of HPV+ tumors from gene expression profiling data.
}
\details{
One or multiple tumor samples to be predicted as HPV+G1 or HPV+G2 based on their expression profiles of the gene panel.
}
\examples{
path = system.file("extdata", "example.txt", package = "PreHPVcc", mustWork = TRUE)
input_data = data_pre(path, type = "ID")
hpv_pre(input_data)
}
\author{
Shengwei Li <lishengwei98@163.com>, Xiaosheng Wang <xiaosheng.wang@cpu.edu.cn>
}
