% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreHPVcc.R
\docType{package}
\name{PreHPVcc-package}
\alias{PreHPVcc-package}
\alias{PreHPVcc}
\title{Prediction of hpv+ tumors based on gene expression profiles in cancer}
\description{
Prediction of hpv+ tumors by the 50 genes
}
