% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_data.R
\docType{data}
\name{training_data}
\alias{training_data}
\title{Training set}
\format{
A dataframe with 281rows (tumor samples) and 51 columns (HPV+ status of tumor samples, and 50 gene features).
}
\source{
\url{https://xenabrowser.net/datapages/}
}
\usage{
training_data
}
\description{
The training set is the expression profiles of a 50 genes
}
\keyword{datasets}
